+++ pkg/modelconv/artifactconv.go
@@ -8,10 +8,11 @@ import (
 // DBArtifactToResponse converts a database artifact to a response artifact.
 func DBArtifactToResponse(dbArtifact database.Artifact) response.Artifact {
 	return response.Artifact{
-		ArtifactID: dbArtifact.ArtifactID,
-		BuildID:    dbArtifact.BuildID,
-		Name:       dbArtifact.Name,
-		FileName:   dbArtifact.FileName,
+		TimeMetadata: DBTimeMetadataToResponse(dbArtifact.TimeMetadata),
+		ArtifactID:   dbArtifact.ArtifactID,
+		BuildID:      dbArtifact.BuildID,
+		Name:         dbArtifact.Name,
+		FileName:     dbArtifact.FileName,
 	}
 }